Splits fmp into two components, finishes initial implementation
