+++ src/common/components/Message/index.js
@@ -3,14 +3,14 @@ import styles from './styles.less';
 import classNames from 'classnames/bind';
 const cx = classNames.bind(styles);
 
-const Message = ({ children, type, title, className, size }) => (
+const Message = ({ children, type, title, className, size }) => (children ? (
   <div
     title={title}
     className={cx('container', className, type, size)}
   >
       {children}
   </div>
-);
+) : null);
 
 Message.propTypes = {
   title: PropTypes.string,
